Handle NO_DOCS response during sane_read calls
The SANE API says that SANE_STATUS_NODOCS can be returned by sane_read.
This happens with scanner drivers that can't detect an ADF out of paper
until the reads begin.

When this happens, the frontend scanner should handle this as a normal
end of documents, and not an error.

When detected now, just clost the document and do not report a warning.