Don't exclude INACTIVE descriptor from list of options
The Sane.Capability.INACTIVE capability indicates that
the option is not currently active (e.g., because it's meaningful
only if another option is set to some other value).

With some proprietary backends (example: esci, Epson WF-7710),
the 'duplex' option is INACTIVE at the start.
To activate it, the 'source' option need to be set to 'ADF' value.

With this patch, check if option is INACTIVE is made only during
value set, and not during creation of list of available options.

Fixes: #106