Remember save-directory based on actual file URI
Derive the "save-directory" setting to save from the
file URI instead of using the
'Gtk.FileChooser.get_current_folder_file()' method in order
to reliably remember the directory that the file was actually
saved to.
This also makes the actual URI be stored instead of the file
path.

In particular, 'Gtk.FileChooser.get_current_folder_file()' does
not return the actual directory when GtkFileChooserNative is used
with a FileChooser portal (as can be tested e.g. with the
'GTK_USE_PORTAL=1' environment setting).