app-window: Use a better crop icon
This replaces the scissors icon, which is usually meant for when you
to cut (i.e. copy and delete) something rather than cropping. There's
a nice crop icon in the icon dev kit so better make use of it so it's
more clear what's the purpose of the crop button.