Allow scan double side of single page if Flatbed is not found
With previous implementation the Scan Single Page
was working only in case the Flatbed is detected.
It seems that for ADF only scanners, scanning Single Page
could be also useful. Unfortunately simple-scan is scanning
only front page. With this commit it is possible to scan
Single Page with both sides on ADF only scanner.

Fixes #188